backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
pillow==11.3.0
sqlparse==0.5.3
django-filter==24.3
